V1.0.0 (#14)

+++ src/TabPanel/TabPanel.hpp
@@ -12,22 +12,21 @@ class TabPanel final : public QDockWidget {
     explicit TabPanel(QWidget* parent = nullptr);
 
     void init();
-    void
-    addTab(const QString& basename, u32 total, u32 translated, bool completed);
-    [[nodiscard]] auto tabCount() const -> u16;
-    [[nodiscard]] auto tabName(u16 tabIndex) const -> QString;
+    void addTab(const QString& basename, i32 total, i32 translated, bool completed);
+    [[nodiscard]] auto tabCount() const -> i32;
+    [[nodiscard]] auto tabName(i32 tabIndex) const -> QString;
     [[nodiscard]] auto currentTabName() const -> QString;
     void clear();
     void changeTab(const QString& filename);
 
-    [[nodiscard]] auto tabIndex(const QString& tabName) const -> u32;
-    [[nodiscard]] auto currentTranslated() const -> u32;
-    [[nodiscard]] auto currentTotal() const -> u32;
-    [[nodiscard]] auto tabTotal(u16 tabIndex) const -> u32;
-    [[nodiscard]] auto tabTranslated(u16 tabIndex) const -> u32;
-    void setTabTranslated(u16 tabIndex, u32 translated) const;
-    void setCurrentTranslated(u32 translated) const;
-    void setCurrentTotal(u32 total) const;
+    [[nodiscard]] auto tabIndex(const QString& tabName) const -> i32;
+    [[nodiscard]] auto currentTranslated() const -> i32;
+    [[nodiscard]] auto currentTotal() const -> i32;
+    [[nodiscard]] auto tabTotal(i32 tabIndex) const -> i32;
+    [[nodiscard]] auto tabTranslated(i32 tabIndex) const -> i32;
+    void setTabTranslated(i32 tabIndex, i32 translated) const;
+    void setCurrentTranslated(i32 translated) const;
+    void setCurrentTotal(i32 total) const;
     void setProgressDisplay(bool percents) const;
 
     [[nodiscard]] auto tabs() const -> QStringList;
